Use preprocessor macros for dimension-specific ljfluid code.
Macros represent the most ugly, but also the most straightforward solution
for keeping 2D and 3D specific code sections to a minimum.
A template parameter <int dimension> with dimension-specific member functions
would be preferable, but seems hard to implement if use of inheritance is to
be avoided at all costs.