Fix extremely stupid data corruption bug.
If one asynchronously copies memory from GPU to host, the relevant CUDA
stream has to be synchronized before accessing the data in host memory!

For the curious, a note on how this bug eventually hit the author on the
front head: After hours and hours of tinkering with the GPU kernel code for
cell updating, the brilliant idea arose to printf the cell placeholder data
in host memory before processing it.  As a side effect, the data corruption
was suddenly gone. The reason: the printf loop took so long to execute that
in the meantime the memcpy operations from GPU to host could finish.