Fix out-of-bounds index in cell computation.
Mapping the positional coordinates of a particle to its corresponding cell
index is the most delicate part of the cell lists update. The safest way
is to combine round-towards-zero with a successive integer modulo operation,
which comes with a performance penalty.

As an efficient alternative, we transform the coordinates to the half-open
unit interval [0.0, 1.0) and multiply with the number of cells per dimension
afterwards.