Fix center of mass velocity drift with double-precision force summation.
The summation of all forces acting on a particle is the most
critical part what concerns longtime accuracy of the simulation.

Naively adding all forces with a single-precision operation is fine
with the Lennard-Jones potential using the N-squared algorithm, as
the force exhibits both a repulsive and an attractive part, and the
particles are more or less in random order. Thus, summing over all
forces comprises negative and positive summands in random order.

With the WCA potential (Weeks-Chandler-Andersen, purely repulsive
part of the shifted Lennard-Jones potential) using the N-squared
algorithm, the center of mass velocity initially set to zero
undergoes a peculiar drift, but finally saturates at |v| ~ 5e-6.
Using the cell algorithm with the WCA potential however results
in a continuously drifting center of mass velocity, independent
of the chosen simulation timestep.

The reason for this behaviour lies in the disadvantageous summing
order: With a purely repulsive potential, the summed forces of a
single neighbour cell will more or less have the same direction.
Thus, when adding the force sums of all neighbour cells, we add
huge force sums which will mostly cancel each other out in an
equilibrated system, giving a small and very inaccurate total
force due to being limited to single-precision floating-point
arithmetic.

Indeed, implemeting the summation with a native-pair arithmetic,

  Seppo Linnainmaa, Software for Doubled-Precision Floating-Point
  Computations, ACM Trans. Math. Softw., ACM, 1981, 7, pp.272-283,

remedies the velocity drift, with fluctuations < 1e-7, as would
be expected from a symplectic integrator such as the Verlet
algorithm used here.