mdsim/gpu/boltzmann: fix repeating random numbers due to same generator state.
While debugging the rare potential energy divergence with a thermostat,
which is *not* fixed in commit 0ec697330cc9ad28730c858fcd88450a0f48692f,
disabling particle permutation in memory due to Hilbert curve ordering
revealed a grave bug in the random number generator: We forgot to store
the generator state back to global memory.

This caused a recurring Boltzmann velocity distribution, seen in
temperature oscillations between the lattice relaxation and the
thermostat temperature.