trajectory: fix weird bug occuring with the HDF5 library on OpenSuSE 10.2.
This bug occurs regardless of wether the binary is compiled and linked
dynamically on the target system (OpenSuSE 10.2), or compiled and linked
statically on another system (OpenSuSE 10.3).

It seems the root node is closed upon overwriting the H5::Group inside
the loop in the case of a bidisperse system with trajectory subgroups
'A' and 'B'. As a workaround, we assign either the root node for a
monodisperse system, or the subgroup for a bidisperse system.


HDF5-DIAG: Error detected in HDF5 library version: 1.6.9 thread 0.  Back trace follows.
  #000: H5G.c line 196 in H5Gcreate(): not a location
    major(01): Function arguments
    minor(03): Inappropriate type
  #001: H5G.c line 1895 in H5G_loc(): invalid group ID
    major(01): Function arguments
    minor(05): Bad value
  #002: H5I.c line 1257 in H5I_find_id(): invalid ID
    major(07): Atom layer
    minor(32): Unable to find atom information (already closed?)
terminate called after throwing an instance of 'H5::GroupIException'
Aborted