ljfluid_gpu_nbr: check for overflowing cell or neighbour lists.
For the cell and neighbour lists updates, we add a global return value
which indicates the occurance of fully occupied cell or neighbour lists.

Prior to this change, particle loss seemed to reliably result in a
potential energy divergence after further steps. However, this did not
allow to differentiate between errors in the cell or neighbour list
updates and other causes such as an excessive time-step. More importantly,
the annihilation and successive creation of a particle might as well
have gone unnoticed in rare cases.