ljfluid_gpu_nbr: fix uninitialized neighbour lists.
Neighbour lists on the GPU are written with a stride size of the number
of particles *rounded up* to the nearest multiple of the block threads
count. The additional memory for placeholder particles is therefore
distributed among the entire length of the neighbour lists vector.

Setting the size of the vector to a smaller value than its effective
capacity causes cuda::memset not to erase neighbour list placeholders at
the end of the vector.