correlation: apply optional --max-samples value*s* to lowest blocks only.
The purpose of a maximum number of samples per block is to mitigate
performance bottlenecks due to high sampling frequencies over the course
of the simulation; thus limiting the number of samples only makes sense
for the lowest block levels. This also allows for optimal statistics in
the intermediate time regime.