ljfluid_host: fix out-of-bounds index in binary mixture trajectory sampling.
This bug was introduced in commit a6287979fad871f486ab2f9a4af68116a9163dca.

A segmentation fault occurred if simulating binary Lennard-Jones mixtures with
the host backend. Note that test and production runs with the host backend of
commit a6287979 or later had only been done for monodisperse systems, which
were not affected by this bug.

The particle tags for each species have to be chosen from within the species's
tag range, which is mapped to the index range of the species's sample upon
sampling the particle trajectories.