Replace CUDA streams and events with host timers to work around long-time freezes.
It was discovered that long-time simulations reproducibly freeze after 10⁸ or
more integration steps. The process continues to consume CPU while effectively
hanging.

The amount of steps is reproducible and does not seem to depend on the initial
trajectory, but varies with system size. For example, a simulation of a
Kob-Andersen mixture with 40,000 A and 10,000 B particles hangs after ~1.1×10⁸
steps, while a simulation of a mixture of 2,000 A and 500 B particles at
the same density and temperature hangs after ~1.7×10⁸ steps.

The freeze occurs in the MD step synchronization, when the last CUDA
event inserted into the MD step stream is synchronized. The issue is not
mitigated by creating and destroying the CUDA events before and after use to
avoid reuse.
If the CUDA events are omitted and the stream is synchronized instead, the
simulation of the 50,000 Kob-Andersen test case goes beyond 1.1×10⁸ steps, but
eventually hangs as well after ~1.3×10⁸ steps.

As a work-around we avoid any use of CUDA streams and events, and use
the high-precision host timer for profiling.
This increases the measurements by a tiny (presumingly constant) amount
due to the inclusion of the CUDA kernel invocation time, and prohibits
concurrent kernel and memory copy operations.

With this fix the simulation of the first test case accomplished 2.0×10⁸ steps.