add top-level options --force and --integrator
These options allow the selection of a specific force module
before module resolution.

Currently only --force=power-law works, with --force=lj
there seems to be a bug in the module logic. When the powerlaw
module is discarded, its dependencies (particle, box) are discarded
as well and not again resolved by the lj module.