Prepare for upcoming module mechanism rewrite.
This commit introduces changes to the module API.

The function resolve() is split into depends() and select(),
separating dependency resolution from the querying of the
suitability of a module. This will be needed to assemble a
module dependency graph before options parsing.

The dependency directives

    module<target_type>::required(vm)
    module<target_type>::optional(vm)

change to

    modules::required<_Self, target_type>()
    modules::optional<_Self, target_type>()

The new typedef _Self points to the module class itself.

The construction directive changes from

    module<target_type>::fetch(vm)

to

    modules::fetch<target_type>(vm)