thermodynamics: fix reference value of numerical precision.
The numerical precision of the simulation backend should either be
determined by querying the backend itself (which is not implemented),
or by using the macro to switch between single and double precision.

This commit might break some test cases, as the lower boundaries for the
precision of the thermodynamic properties, which were previously a mixture
of single and double precision epsilons, now reflect the precision of
the backend.