consistently rename virial tensor to stress tensor
The virial sum or function is the sum of the diagonal elements of the
potential part of the stress tensor, it is not really a tensor by
itself.

The division after particle types has been dropped since it introduces
many questions about physical definitions and makes the programmer's
life more difficult. And it is not clear whether somebody ever will
need it.