introduce mdsim::type_traits
collect various definitions of vector types scattered over the source.

This is mainly used for primary definitions of vector and tensor
entities with the aim of reducing references to fixed_vector and of
making all boost::mpl::if constructs for float4/float3/float2 etc.
consistent.

A module depending on another module should still use the definitions
provided there.