Remove option group descriptions.
As the option namespace is global, option names should be chosen wisely
to give the user a hint to which module(s) it belongs. Proper naming
should not be substituted with a quickly chosen group name.

This also limits the clutter in the option summary, which serves as a
quick reference to the user, instead of a detailed description in the
documentation or manual page.