Use __attribute__((constructor)) for static initialization.
This is easier to comprehend than static dummy variables.

Note that __attribute__((constructor)) is not portable, but works with GCC,
Clang (LLVM) and icc, which should cover all C++ compilers on HPC systems.

Neither is the -whole-archive ld flag required with static linking.

The double parentheses are needed for namespace_ concatenation with comma.