Fix Lua C++ class registration with GCC 4.3 and 4.4.
Pass dependency injection unit test with GCC 4.3 and 4.4.

The way we used luabind::scope causes havoc with GCC >= 4.3.

Previously every Lua registration function returned an instance of
luabind::scope, which was stored *as a copy* in an STL container.
The registration with the Lua state (L) was then done by passing each
scope in the registry to module(L). If compiled with GCC 4.3 or 4.4,
this method of class registration resulted in cast failures and
unregistered class exceptions.

To fix this issue we use the normal registration procedure as found
in the Luabind documentation, which avoids copies of luabind::scope.
Instead of scopes the registry stores registration functors, which
upon registration are invoked with the Lua state and use module(L).