gpu::forces::lj: add GPU kernel wrapper for pair_short_ranged.
We want to keep the interfaces of the pair force loop and the potential
separate so a change in the compute() function signature does not require
changes to the potentials. Instead of wrapping the compute() function and
dependent constants, we therefore add a template wrapper for pair_short_ranged
that is instantiated for each potential.