numeric/blas: implement Boost Collection concept in GPU storage array for fixed_vector.
Currently fixed_array is defined differently on GPU and host.

On the GPU we use a custom implementation, whereas on the host we derive
from boost::array. In particular, the GPU implementation contains a hack
which replaces the element data type "dsfloat" with "float2", which was
needed because CUDA 2.3 does not allow __shared__ structs containing
data members that define a constructor.

This hack causes problems with alignment of kernel parameters, which in
the case of fixed_vector<dsfloat, …> may differ between GPU and host.

For CUDA 3.1 and 3.2, this commit introduces a proper fixed_array that
contains a C array with the correct data type. Further, fixed_array now
models a Boost Collection, which is needed later to unify GPU and host
implementation and drop boost::array.

This commit refs #55.