rewrite unit test for position::lattice
The generated lattice is tested by computing its structure factor for
some wavenumbers of the reciprocal lattice. The previous version of
the test compared absolute positions between host and GPU
implementation, which is a delicate matter due to different particle
order.

The test is still not able to detect the bug fixed by commit b1f4f8f.