mdsim/forces: add pair_trunc Lua module.
We have separate C++ modules for pair force and potentials, so the Lua modules
should be separate as well, to avoid redundant code for each potential.

The separation further exposes the potential modules to the hooks registry,
which is needed in a later commit to implement write_parameters for the
potential parameters.

Remove redundant C++ pair_trunc<power_law> wrapper.