version.h: fix generation of <halmd/version.h> in pristine build tree.
This fixes a regression from commit 23512f0, which moved options parser to a
subdirectory, which breaks the CMake rule for generating version.h as CMake
rules do not propagate across directories.

Instead of moving the options parser back to the main directory, we strip the
options parser of the functionality *not* related to parsing: The main program
options (e.g. --version, --help) are defined as well as handled in main().
This removes the dependency of the options parser on <halmd/version.h>.