cmake: install Lua script files into HALMD installation prefix.
This commit fixes #51.

Copy Lua scripts from source to build tree, with @variable@ substitution.

Extend package.path following principle of least surprise: The default
paths are searched first, which may be overridden by the user with the
LUA_PATH environment variable. The *relative* path to the Lua scripts
directory in the build tree is searched next, which is only valid if
the halmd binary is executed from its build subdirectory. Finally, we
search the CMake installation prefix.