logger: fix FTBFS with Boost.Log > 1.0.
The attribute interface of Boost.Log changed in SVN r479.
Boost.Log does not provide any version information, but
we may assume that users of Boost >= 1.46.0 require at
least Boost.Log r588 to compile with Boost Filesystem
version 3. For users of Boost < 1.46.0 we assume that
Boost.Log 1.0 was used as recommended in the HALMD
prerequisites guide.