density_mode: take GPU trajectory sample as input
This avoids costly data transfer and allows computation of density
modes on the GPU.

Since trajectory sampling on the GPU is not yet implemented, we resort
directly to the positions in the particle module and support only a
single particle type so far.

The corresponding Lua script has been broken,
halmd.observables.trajectory requires some way to select between host
and GPU samples.