trajectory/writer: implement on_append signal.
Add signal which is emitted before appending the trajectory. This will
be used to register phase_space::acquire as a callback and invoke it as
needed, i.e. following the sampling interval of the trajectory writer.

To prevent forgotten callbacks due to scripting errors, validate the
time of the phase space sample against the current simulation time.