observables/ssf.lua: connect ssf::on_sample signal to density_mode::acquire slot.
The ssf module already contains an instance of density_mode and thus
could simply call density_mode::acquire before sampling, as before.

However we want to be consistent with program design, which means that
the sampling control flow is setup entirely in Lua using signals.

Note the positive effect on the ssf unit test: density_mode->acquire has
to be called explictely. If for example density_mode::acquire threw an
exception, the stack trace would not contain the ssf::sample method,
which is consistent as it is not involved in the computation of
density modes.