sampler: migrate to signal-based sampling.
Add two signals on_prepare and on_sample, which are emitted before and
after preparation of the simulation, and before and after an MD step.

In the Lua script, observable::prepare and observable::sample are passed
to the sampler as a pair to avoid forgetting observable::prepare, along
with the desired sampling interval.

Move sampler interval options (currently --state-vars and --trajectory)
to the “halmd” module, and make each module optional. For performance
tests one might not want to sample any observables at all.