Revert "Revert "cmake: explicitly enable only C++ language.""
This is the second attempt to enable only the CXX language.

It turns out that CMake has a bug in find_library which ignores lib64 paths if
the C language is not enabled. The bug in find_library has been reported[1]
upstream, along with a patch[2] that fixes the problem.

[1] http://public.kitware.com/Bug/view.php?id=10813#c25824

[2] http://public.kitware.com/Bug/file_download.php?file_id=3759&type=bug