lattice: shuffle positions instead of types
The particles' tag and type are not independent. For example, acquisition of a
phase phase sample relies on a continuous assignment of tags across all
particle types. (This might be changed in the future.) Instead of shuffling
particle types and tags together, we simply shuffle the generated lattice
positions.

The latter approach was implemented for the GPU version already. The operation,
however, took place before assignment of the lattice points. This has been
fixed.

Inconsistent shuffling of particle types resulted in corrupted memory accesses
upon acquisition of the phase space sample. The segmentation violation,
however, appeared much later during HDF5 output of the trajectory.

This refs #45.