Revert "Use static struct instead of anonymous namespace in HALMD_INIT."
The effect of static struct is in fact *not* the same as of an anonymous
namespace. The static keyword only reduces the visibility of the variable
to the translation unit (e.g. _register_luaopen), but not the visibility
of the class constructor (register_luaopen::register_luaopen).
This results in duplicate symbol errors at the linking stage.

This reverts commit 072653ca022936f863119d4d29455c1fb95c106d.