Explicitly connect thermodynamics::prepare to sampler::on_prepare.
Simplify sampler interface by removing slot pair function type.

The goal is to keep the sampler interface simple to use and consistent,
therefore it is better to connect prepare and sample in two steps. It
is the sampled module's responsibility to alert the user of a missing
signal connection, e.g. to <module>::prepare. This is already the case
with thermodynamics, which throws an exception on sample if prepare
has not been called before.