utility/lua: implement portable boost::any to Lua converter.
The previous implementation used &typeid(T) to store a type T as
a key in the any converter map. This is not portable (e.g. to
AIX), as there is no guarantee that &typeid(T) == &typeid(T) for
identical type T. Instead we use the Luabind class type_id, which
uses the well-defined method type_info::before to compare types.