cmake: use one module per target to minimize unit test link dependencies.
The GNU linker has no problem with dropping unneeded objects from a
linked archive. The AIX linker, however, uses the *entire* archive
and thus needs link dependencies for all objects, which causes many
missing link dependencies for the unit tests.

http://unix.derkeiler.com/Newsgroups/comp.unix.aix/2003-08/0081.html

There are two solutions to this problem:

Link everything against everything. This is a bad solution, as a change
in a single module would trigger re-linking of the entire unit tests.

Make targets more granular, use one module per target. This is the
better solution, and minimizes the link dependencies needed on AIX.
Tests are re-linked only when their dependent modules are updated.

Obviously the number of targets no longer scales with the number of
directories, but the number of modules. The only remedy is parallel
compilation with make -j to avoid waiting for I/O.