cmake: the magic AIX linker options are -qtwolink -Wl,-bweaklocal -qnoipa.
The combined flags -qtwolink -Wl,-bweaklocal instruct the linker to discard
unreferenced objects in an archive, which is needed to link unit tests with
minimal link libraries. -qnoipa is needed to disable link-time IPA, which
breaks -qtwolink. This way we retain the single target per directory layout,
which is much better in terms of build I/O.