cmake: compare CMAKE_* with CMAKE_*_INIT to check for CMake defaults.
This is the safest way to determine whether flags have not been set by
the user, and may be overriden with our defaults.
Note that we cannot set the CMAKE_*_INIT flags directly, as these are
not cache variables and may be set by CMake compiler modules. Further,
we don't know the compiler and platform before enabling the compiler
languages, and afterwards it is too late to set initial default flags.