Define halmd_add_library and halmd_add_modules macros.
The halmd_add_library macro is a wrapper around add_library, which further
adds the library target name to a global property list of HALMD libraries.
The HALMD executable is linked against these libraries.

The halmd_add_modules macro adds all arguments to a global property list of
HALMD modules. This list is used to generate calls to luaopen_<module>
functions, which exports a C++ module into the Lua interpreter.