Revert "test/tools: add shared library test_tools_logger with logger fixture."
Putting the logger fixture into a shared library is not as easy as it
seems. The logger has a Lua binding function, which requires it to be
linked against the Luabind library. On Linux/x86_64, if the Luabind
library is statically linked, as recommended in the prerequisites guide,
this requires cxxflags=-fPIC on compilation of Luabind to make the
library relocatable.

On AIX/Power6, executables linked against libtest_tools_logger fail:

Could not load program test_unit_algorithm_host_pick_lat�c�:
rtld: 0712-001 Symbol decimal128ToNumber was referenced
      from module test/tools/libtest_tools_logger.so(), but a runtime definition
            of the symbol was not found.
rtld: 0712-001 Symbol decContextDefault was referenced
      from module test/tools/libtest_tools_logger.so(), but a runtime definition
            of the symbol was not found.
rtld: 0712-001 Symbol decNumberCopy was referenced
      from module test/tools/libtest_tools_logger.so(), but a runtime definition
            of the symbol was not found.
…

Until these issues are solved, revert the logger fixture.

This reverts commit 0a761999b2c2d6c9fb9844353f5e77d05614012d.