singleton.lua: store strong references to C++ singleton objects.
C++ singletons were previously stored as *weak* table values, which
means a C++ instance is deconstructed when the reference count of the
shared_ptr holding the instance goes to zero. This caused the device
singleton to be destroyed upon trajectory continuation, as no module
explicitly holds the device instance in this case.

To fix this issue, use strong references in halmd.singleton. At a later
stage halmd.module will be rewritten to allow weak and strong singletons.

This fixes #69.