gpu::particle: move allocation of double-single precision memory from verlet to particle.
Double-single precision on the GPU requires additional memory for
particle position and velocity coordinates, which was previously
allocated in the Verlet (and Verlet-Andersen) integrator.

This was admittedly bad style, and could lead to bugs if one forgot
to allocate the additional memory in a newly implemented integrator.

Therefore we allocate the memory once only, in the particle c'tor.