Revert "main: revert to not catching expections if compiled in debug mode."
This reverts commit 0243579d588dcbb34b9173961282e18b0174609e.

Not catching exceptions is bad practice, regardless of whether the
program is compiled for debugging or not. Unhandled exceptions are
easier to catch with a debugger, but cause the program to terminate
without proper deconstruction of objects. In the case of the HDF5
trajectory writer, this will leave the HDF5 file in an inconsistent
state, which in fact makes debugging harder.

Conflicts:

	halmd/main.cpp