mdsim::core: fix microscopic state preparation.
The core::on_prepend_force signal only rebuilds the neighbour lists if
the maximum squared displacement criterion evaluates to true.

Add a separate set of on_*_prepare signals to core, where on_prepare
sets the positions and velocities, and on_append_prepare rebuilds the
neighbour lists and computes the initial forces.

Note that this design is not yet optimal, as the position and velocity
modules are unnecessarily kept alive during the simulation. On the other
hand, the profiler has a design flaw that causes memory corruption if
profiled modules are deallocated before the profiler, so for now it is
in fact necessary that position and velocity remain allocated.