MSD: let Lua module return an appropriate phase space sample
The TCF functor is the place where the input data of the correlation are
defined. In case of the MSD, this is simply a phase space sample, but could
also be a density_mode sample in case of a scattering function. The calling
correlation module will construct a sample::blocking_scheme out of the returned
sample.