Revert "utility/signal: implement signal proxy."
This reverts commit eec46dff2dfaacf66b3d987e0a26639fe026aecc.

The signal proxy is in fact not very useful, as it requires both the
slot iterator (signal::connection) and the signal or signal proxy to
disconnect a slot. A better solution to disconnect a slot would be a
connection object returned on signal::connect that contains the slot
iterator and a reference to the signal.