Disentangle gpu::particle and device modules.
The number of threads per block is an implemention-specific parameter
of gpu::particle (which propagates to other mdsim::gpu modules), and
not a generic parameter of device.

Moving the threads parameter to gpu::particle, and adding a default
value to the constructor, gives gpu::particle and host::particle a
common constructor. This eases instantiation in unit tests.

Move device from halmd::utility::gpu to halmd namespace, to allow
convenient use in all gpu modules. Move validation of the number
of threads to static method device::validate.