device: replace “CUDA device” with “GPU” as prefix of logged properties.
“GPU” is shorter, and independent of the GPU driver/library name.

Remove the device::properties attribute, as modules are not supposed to
query the device instance, but rather query the device properties using
the CUDA runtime. This keeps the constructors of gpu and host modules in
sync, which eases implementation of unit tests.

Remove major and minor “revision”, this is the CUDA compute capability.