host::neighbour: support multiple particle and binning instances.
This is highly experimental. It is assumed that the binning instances
are correctly setup with the same parameters, and only a single species
spanning all particle instances is allowed.

This commit changes the definition of particle pair permutations within
a single instance: Instead of comparing the type-dependent tags, we
directly compare the particle indices. The indices are equally suited
to identify particles as long as the particles are not reordered, which
requires re-binnning anyway.