host::samples::phase_space: add getter and setter slots.
A getter slot allows writing the particle positions or velocities to a
HDF5 or other dataset using the upcoming H5MD writer. The getter slot
is passed from the phase_space sample to the writer.

A setter slot allows reading the particle positions or velocities from a
HDF5 or other dataset. The setter slot is passed from the reader to the
the phase_space sample.

For completeness, allow setting particle positions or velocities by copy
from a sample vector, e.g. using a Lua table of tables with coordinates
for each particle.