mdsim::box: add getter method for edge vectors.
The edge vectors ((L_x, 0, 0), (0, L_y, 0), (0, 0, L_y)) are needed
for the edges dataset in the box group of the H5MD trajectory group.
Note that this feature of H5MD is still in discussion.

Wrap origin and edges as data slots, which allows writing their
values as a time series for use with a NPT ensemble integrator.

Move volume, origin and edges from header to source file. These members
are not used in performance critical loops, and further computed on the
fly instead of stored as a private member variable.

Declare protected class members as private.