script: always print Lua stack trace on error.
Whether compiled in debug mode or not, a Lua stack trace greatly
simplifies the search for the source of a C++ exception, e.g. if
an invalid argument is passed to C++ method in a Lua module.

Set Luabind's lua_pcall error handler in script constructor.

Remove unneeded scoped error handler.