host::samples::phase_space: implement proper argument validation.
assert() is not suitable for argument validation, as assert statements
are discarded if compiled without debugging (if NDEBUG is defined).

Throw std::invalid_argument exceptions with helpful descriptions.

Note that the error messages are prefixed with "Invalid argument: "
by the Lua exception handlers registered in the script class, so
the word "invalid" may be omitted from the error strings.