halmd::signal: implement connection object.
Return a connection object upon connecting a slot to the signal,
which allows later disconnection of the slot. This is similar
to the functionality in Boost.Signals, although our connection
class is a simple template with the slot function signature.

As the connection objects will be used to disconnect slots in Lua,
their usability is not affected by the template implementation.