Implement signal connection tracking and disconnection in Lua.
This module implements connection tracking for a C++ module that provides
signals. When a slot is connected to a signal, a connection object is
returned, which allows later disconnection of the slot from the signal.
With a raw C++ signal method, a connection of a slot looks like this:

  core:on_prepend_force(neighbour.update)

If a Lua module is patched with signals.patch, this turns into:

  core:on_prepend_force{neighbour.update, tracker = neighbour}

i.e. neighbour becomes the tracker of the connection of the slot
neighbour.update to the signal on_prepend_force. This allows
disconnecting the slot later on, provided that module (in this
example core) still exists:

  neighbour:disconnect()

A tracker may track an arbitrary number of signal connections.
Further the tracked connections may be moved to another module:

  neighbour:reconnect(force)

This allows tracking connections of different modules with a single
module, with disconnection of a module along with its dependencies:

  force:disconnect() -- disconnects neighbour.update slot