io::writers::h5md: add H5MD truncate mode writer.
This module implements collective writing to one or multiple H5MD
datasets. Upon initialisation, the writer is assigned a collective
H5MD group. A dataset within this group is created by connecting a
data slot to the on_write signal.

The writer provides a common write slot, which may be connected to
the sampler to write to the datasets at a fixed interval. Further
signals on_prepend_write and on_append_write are provided to call
arbitrary slots before and after writing.