signal: support duplicate disconnect of same slot.
After erasing the slot using the iterator stored in the connection
object, reset the weak_ptr, which breaks the connection to the
shared_ptr and causes weak_ptr::lock to return an empty pointer
on the next call to connection::disconnect().

In signal::disconnect_all_slots, to signal the connection object
that all slots have been removed, reallocate the slots list, which
again breaks the connection of the weak_ptr to the shared_ptr.