Separate logger source from logger class.
Rename halmd::logger class to halmd::logging to free up the name
halmd::logger for the logger source itself. This class implements
a logging interface with severity levels. More accurately it adds
a console and a file sink to the log core, and supports discarding
messages below a desired severity level.

We use the Boost.Log library, which differentiates between log
sources and log sinks. The logging class handles only log sinks,
and therefore is needed only to setup logging console and file.

We provide a default logger source for use with the LOG* macros.
For maximum convenience, this logger is declared as a variable
instead of function. If a module should use its own logger, e.g.
to add attributes such as the module name, one may conveniently
declare a class member logger_, which overrides halmd::logger_.

As module loggers will be created in Lua to automatically add
attributes such as the module name, we declare logger_ as a
shared_ptr, which allows passing it to the module constructor
(loggers have no copy constructor). As a safety measure we
declare the shared_ptr as const, while the logger itself is
mutable.

Move log formatting to shared private method, and switch to
lambda-style formatting with support for conditional output.

Output severity level as integer for unknown severity levels,
as implemented for the trivial severity logger in Boost.Log.

Reuse HALMD_LOG macro in HALMD_LOG_ONCE.

Drop support for Boost.Log < SVN r479 (e.g. version 1.0). As
we require Boost 1.43.0, these older versions need no longer
be supported.

Rename Luabind wrapper for HALMD_LOG from log_wrapper to wrap_log,
which is consistent with the names of other Luabind wrappers.