modules.lua: provide module logger as args.logger.
To make it easy to separate log output of different modules, support
optional log attributes Module and Tag, which will be output as log
line prefix "Module: " or "Module[Tag]:".

Pass a newly constructed logger to module constructors as args.logger,
with the Module attribute set to the module namespace. A module may
opt to further add the Tag attribute, or discard the logger if not
required.