box: add extend_periodic()
This function is the counterpart of reduce_periodic(). For the moment, it is
only used by the phase space samplers. A centralised function shall ensure that
the behaviour is consistent throughout the program and that it actually
provides the inverse operation of reduce_periodic. The latter should be checked
by a unit test.