fix test box/periodic/gpu/3d
Reallocating memory screwed up the previously copied data.

We extend the kernel by explicitly passing the size of the data vector, which
avoids the need of memory reallocation.

For large box sizes, the tolerances on the reduced positions had to be
adjusted. Thus, the periodic reduction is not strict due to roundoff errors.