utility.profiler: implement runtime accumulator registration in Lua.
Register module object hook which invokes module.profile if defined.
Pass a module profiler to this function, with profiler:on_profile to
register runtime accumulators of a module object. profiler:on_profile
takes a single unnamed argument, which is the module to be profiled,
and arbitrary many named arguments, with each the name of a runtime
accumulator and its description.

Track accumulator connections with halmd.modules.signals.

Connect profiler.profile to sampler:on_finish.