utility/signal: implement slot-type-independent connection objects.
Wrap the typed slots<T>::connection_ object with a boost::function
object, which allows modules to use the same connection type with
any type of signal or slots container.

This removes the implementation-specific slot type from the public
API of a module. Further the connection object is a slot itself,
so it may be connected to another signal to disconnect the slot.