writers::h5md: pass root group as constructor argument instead of file.
There is no need to require passing a file reference (or shared_ptr).
This change decouples the h5md::file and append/truncate modules.

HDF5 will keep the file open as long as it is not explicitely closed,
and there remain opened HDF5 objects such as H5::Group or H5::DataSet.