writers::h5md: explicitly use boost::function in on_write function signature.
We actually only support boost::function functors, so this should be
obvious from the function signature of on_write.

This avoids surprises if a boost::bind object is passed e.g. in the unit
tests, which must be converted to a boost::function object before it can
be used within the boost::bind call for write_dataset.
A user of the C++ writer API now has to explicitly specify the data type
along with on_write (e.g. on_write<std::vector<double> const&>), which
will convert a passed boost::bind object to the correct boost::function
object.