h5md::file: write and read H5MD author attribute.
The author attribute contains the real name of the user.

halmd::realname() retrieves the password file entry for the real user id
of the calling process (or optional argument uid). If the GECOS field is
empty, the login name of the user is returned. If the GECOS field is
non-empty, either the part up to the first comma, or otherwise the
entire field, is returned.